Move node code to library as well
